+++ src/app/shared/models/country.model.ts
@@ -1,0 +1,10 @@
+/**
+ * (c) Copyright by Abraxas Informatik AG
+ *
+ * For license information see LICENSE file.
+ */
+
+export interface CountryModel {
+  isoId: string;
+  description: string;
+}